requirements.txt
torch==2.1.0
torchaudio==2.1.0
pyannote.audio==3.1.1
datasets==2.19.0
soundfile==0.12.1
numpy==1.26.4
matplotlib==3.8.0
huggingface_hub==0.23.0
librosa==0.10.1